Update 01/06/2021
In this update there were big changes, it can be said that the module for the management of subjects and groups (add subject and add students to each group) is already finished.
It is still pending:
Add functions to print the ballots
It remains as an idea to be able to add the module of teachers, so that these are assigned to a group, this with the purpose of showing the name of the teacher in the PDF of the group.